+++ bingads/bulk/entities/targets/__init__.py
@@ -1,5 +1,0 @@
-__author__ = 'Bing Ads SDK Team'
-__email__ = 'bing_ads_sdk@microsoft.com'
-
-from .bulk_targets import *
-